backend/pyproject.toml
[project]
name = "job360"
version = "1.0.0"
description = "Automated UK job search system supporting any professional domain"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "pandas>=2.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "rich>=13.0.0",
    "humanize>=4.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "google-generativeai>=0.8.0",
    "groq>=0.11.0",
    "cerebras-cloud-sdk>=1.0.0",
    # Batch 2 — multi-user delivery layer (auth, channels, notifications).
    "argon2-cffi>=23.1.0",          # password hashing (argon2id, OWASP-recommended)
    "itsdangerous>=2.2.0",          # signed session cookies
    "apprise>=1.7.0",               # multi-channel notification sending
    "email-validator>=2.1.0",       # pydantic EmailStr (via pydantic[email])
    "cryptography>=42.0.0",         # Fernet for user_channels credential storage
    # Pillar 2 Batch 2.10 — four-layer dedup. Both are small + C-backed;
    # imported lazily so they only activate when the RapidFuzz / TF-IDF
    # layers are actually invoked (tests that only touch Layer 1 still run
    # without these installed).
    "rapidfuzz>=3.0",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aioresponses>=0.7.0",
    "fpdf2>=2.7.0",
    # Batch 3.5.4 — test-order flake hunter. Loaded on demand via
    # `pytest -p randomly --randomly-seed=<n>`; do NOT enable by
    # default. Default CI runs use `-p no:randomly` to preserve
    # deterministic ordering for the green-baseline contract. Use
    # during cleanup batches only, once a run shows 0f across N
    # seeds it's safe to flip CI over.
    "pytest-randomly>=4.0",
    # Step-0 pre-flight (Tier B) — parallel test execution via
    # `pytest -n auto`. Off by default; opt-in during long-running
    # cycles to cut wall-clock time on the 600+ test suite.
    "pytest-xdist>=3.0.0",
    "mypy>=1.10.0",
]
indeed = [
    "python-jobspy",
]
# Pillar 2 Batch 2.6 — semantic stack (renamed from ``[esco]`` so one
# install path covers BOTH ESCO skill canonicalisation AND Batch-2.6 job
# embeddings + ChromaDB vector search). Heavy deps (~300 MB for
# sentence-transformers + torch CPU wheel, plus ChromaDB); keep optional.
# Enable with ``pip install '.[semantic]'`` then run
# ``python scripts/build_esco_index.py --esco-csv skills_en.csv``
# to generate ``backend/data/esco/{labels.json,embeddings.npy}``.
semantic = [
    "sentence-transformers>=2.2",
    "numpy>=1.24",
    "chromadb>=0.5",
]
# Deprecated alias — kept for compatibility with any operator workflow
# still referencing the old extra. Points at the same packages.
esco = [
    "sentence-transformers>=2.2",
    "numpy>=1.24",
]

[project.scripts]
job360 = "src.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
# Batch 3.5.4 — disable pytest-randomly by default. When the plugin
# is installed (dev dep) it auto-activates; the CI contract is
# "deterministic green baseline", so we opt out unless the developer
# explicitly enables it via `pytest -p randomly ...` during a
# flake-hunt session.
addopts = "-p no:randomly"
# Step-0 pre-flight (Tier B) — registered markers. Keep this list
# alphabetised; custom markers without registration raise
# PytestUnknownMarkWarning and poison CI logs.
markers = [
    "fast: smoke subset runnable in < 5s",
]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "G"]
# UP007 (Optional → X | None) requires py3.10+ at runtime; project supports
# py3.9 so keep the explicit Optional[] form in source.
ignore = ["S101", "UP007"]

[tool.ruff.lint.per-file-ignores]
# Dev CLI helpers: subprocess + SQL-string concat are expected; input comes
# from argparse + known-safe table names. Globs are relative to this
# pyproject.toml (the `backend/` directory), so `scripts/*` matches
# `backend/scripts/*.py` at the repo level.
"scripts/*" = ["S603", "S607", "S608"]
"scripts/bootstrap_dev.py" = ["S105", "S603", "S607"]
"migrations/runner.py" = ["S603", "S607"]

[tool.mypy]
# Step-0 pre-flight (Tier C) — mypy strict gate. We set python_version to
# 3.10 here (not 3.9 despite requires-python) because the codebase already
# uses 3.10+ union syntax (X | Y) pervasively; setting 3.9 produced 62
# bogus "syntax requires 3.10" errors. Runtime support for 3.9 is preserved
# by `from __future__ import annotations` at the top of modules that need
# to run on 3.9 — that's a runtime concern, not a static-check one.
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "backend"

# Ignore missing stubs for heavy / optional dependencies. Many of these
# are lazy-imported at call sites (CLAUDE.md rules #11, #16) and do not
# ship type stubs. Silence them globally so strict-mode doesn't drown
# in "no module named X" / "no type hints for X" noise.
[[tool.mypy.overrides]]
module = [
    "sentence_transformers.*",
    "sentence_transformers",
    "chromadb.*",
    "chromadb",
    "rapidfuzz.*",
    "rapidfuzz",
    "sklearn.*",
    "sklearn",
    "apprise.*",
    "apprise",
    "fpdf.*",
    "fpdf",
    "pdfplumber.*",
    "pdfplumber",
    "docx.*",
    "docx",
    "jobspy.*",
    "jobspy",
    "groq.*",
    "groq",
    "cerebras.*",
    "cerebras",
    "google.generativeai.*",
    "google.generativeai",
    "google.*",
    "google",
    "argon2.*",
    "argon2",
    "aioresponses.*",
    "aioresponses",
    "arq.*",
    "arq",
    "torch.*",
    "torch",
]
ignore_missing_imports = true
follow_imports = "skip"

# Relax strict-mode checks on test code — tests intentionally monkeypatch
# and type-cast widely.
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
